refactor: convert ChatGPT import from CLI command to skill
Move the import functionality from a `sessions import` CLI subcommand
to a bundled skill (`/chatgpt-import`) so the assistant can
interactively walk the user through the import process.
