+++ assistant/src/config/bundled-skills/chatgpt-import/TOOLS.json
@@ -1,0 +1,27 @@
+{
+  "version": 1,
+  "tools": [
+    {
+      "name": "chatgpt_import",
+      "description": "Import ChatGPT conversation history from an exported conversations.json file or ZIP archive. Use dry_run first to preview what will be imported, then run again without dry_run to perform the import.",
+      "category": "import",
+      "risk": "medium",
+      "input_schema": {
+        "type": "object",
+        "properties": {
+          "file_path": {
+            "type": "string",
+            "description": "Absolute path to the conversations.json file or ChatGPT export ZIP file"
+          },
+          "dry_run": {
+            "type": "boolean",
+            "description": "If true, only preview what would be imported without actually importing. Defaults to false."
+          }
+        },
+        "required": ["file_path"]
+      },
+      "executor": "tools/chatgpt-import.ts",
+      "execution_target": "host"
+    }
+  ]
+}
